services/video-streaming/requirements.txt
flask==2.0.2
opencv-python-headless==4.5.5.62
pyzmq==22.3.0
